Send link to verify email
Added table for verification token
Added email verification view to admin panel for user
Added post save for email verification table. It sends email to the user that requested account
If user is_approved in UserProfile (is affiliated with a university) and verified email, then their account is automatically activated. If not, email is sent to us. For now sender and recipient is the same. Will change recipient to serve@scilifelab.se

Removed post save step for UserProfile in studio/views.py
	modified:   common/admin.py
	modified:   common/forms.py
	new file:   common/migrations/0002_emailverificationtable.py
	modified:   common/models.py
	modified:   common/signals.py
	modified:   common/urls.py
	modified:   common/views.py
	modified:   studio/views.py
	new file:   templates/registration/verify.html