Fix SS-1481, Unable to create new MLFlow app instance (#351) (#352)
Co-authored-by: Johan Alfredéen <johan.alfredeen@gmail.com>
Co-authored-by: Hamza <hamza.saeed@scilifelab.uu.se>

Source: https://scilifelab.atlassian.net/browse/SS-1481

The bug was a TypeError: Object of type set is not JSON serializable occurring when saving a Django model instance with a JSONField. The root cause was a Python set literal, {'--gunicorn-opts="--timeout=360"'} (missing a colon), being used for the extraArgs value in the MLFlowInstance.get_k8s_values() method.

The fix involved changing this incorrect set literal to a JSON-serializable list of strings: ['--gunicorn-opts="--timeout=360"'].

Verification was performed by checking the values.yaml of the bitnamicharts/mlflow Helm chart ("oci://registry-1.docker.io/bitnamicharts/mlflow:2.5.2"), which confirmed that extraArgs: [] expects a list of strings, validating the chosen fix for compatibility with the Helm deployment.